requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
discord.py==1.4.1
dnspython==2.0.0
future==0.18.2
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
oauth2client==4.1.3
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynacl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
uritemplate==3.0.1
urllib3==1.25.10
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
youtube-dl==2020.9.20